OO-3662: double check that the test is really a qtiworks test
To retrieve the min. score, max. score and the cut value, double check
if the resource is really based on QtiWorks otherwise fallback to the
old configuration. Fix an issue in statistics where the max. score and
cut value are retrieve from the wrong place. Cache the min. score and
max. score in the results list of the course assessed identities to
prevent reloading the values for every user.