OO-4974: don't persist cluster infos events (and add a time limit)
For large instance, the size of a cluster infos event is quiet large
(1MB) and is sent every second. If these events are queued, it can takes
a lot of RAM. The time limit ensure that such events don't stay in queue
if they are not processed quickly, as they are all the same, one in the
queue is enough.