OO-5124: process create/update to prevent ownership of meeting
If "On behalf" user is configured, the creation of the meeting happens
in to steps, first create the meeting with the communications API to
prevent on "On behalf" user to be the only owner of the meeting and the
only one to use group rooms and to configure the meeting in the Teams
application. In a second step, use the "On behalf" user to update the
meeting with the settings that are not set in the first step.